reorder compilation of let statements
Before the name of the let statement was added first,
if the name was defined before the name in the expression would refer
to the uninitialized new variable

the code:
```
let x = 0
let x = add(1,x)
```
would first eval to an error because the x in add(1,x) refers to an
empty field and not to the x=0.

now it compiles to x=1.